+++ src/flowchart.symbol.js
@@ -4,18 +4,31 @@ function Symbol(chart, options, symbol) {
   this.connectedTo = [];
   this.symbolType = options.symbolType;
 
+  this.next_direction = options.next && options['direction_next'] ? options['direction_next'] : undefined;
+
   this.text = this.chart.paper.text(0, 0, options.text);
+  //Raphael does not support the svg group tag so setting the text node id to the symbol node id plus t
+  if (options.key) { this.text.node.id = options.key + 't'; }
   this.text.attr({
     'text-anchor': 'start',
     'font-size': (this.chart.options.symbols[this.symbolType]['font-size'] || this.chart.options['font-size']),
     'x': (this.chart.options.symbols[this.symbolType]['text-margin'] || this.chart.options['text-margin']),
     stroke: (this.chart.options.symbols[this.symbolType]['font-color'] || this.chart.options['font-color'])
   });
+
+  var font = (this.chart.options.symbols[this.symbolType]['font'] || this.chart.options['font']);
+  var fontF = (this.chart.options.symbols[this.symbolType]['font-family'] || this.chart.options['font-family']);
+  var fontW = (this.chart.options.symbols[this.symbolType]['font-weight'] || this.chart.options['font-weight']);
+
+  if (font) this.text.attr({ 'font': font });
+  if (fontF) this.text.attr({ 'font-family': fontF });
+  if (fontW) this.text.attr({ 'font-weight': fontW });
+
   if (options.link) { this.text.attr('href', options.link); }
   if (options.target) { this.text.attr('target', options.target); }
-  if (chart.options.maxWidth) {
+  if (this.chart.options.symbols[this.symbolType]['maxWidth'] || this.chart.options['maxWidth']) {
     // using this approach: http://stackoverflow.com/a/3153457/22466
-    var maxWidth = this.chart.options.symbols[this.symbolType].maxWidth || this.chart.options.maxWidth;
+    var maxWidth = this.chart.options.symbols[this.symbolType]['maxWidth'] || this.chart.options['maxWidth'];
     var words = options.text.split(' ');
     var tempText = "";
     for (var i=0, ii=words.length; i<ii; i++) {
@@ -41,6 +54,7 @@ function Symbol(chart, options, symbol) {
     });
     if (options.link) { symbol.attr('href', options.link); }
     if (options.target) { symbol.attr('target', options.target); }
+    if (options.key) { symbol.node.id = options.key; }
 
     this.group.push(symbol);
     symbol.insertBefore(this.text);
@@ -115,22 +129,65 @@ Symbol.prototype.getRight = function() {
 
 Symbol.prototype.render = function() {
   if (this.next) {
-    var bottomPoint = this.getBottom();
-    var topPoint = this.next.getTop();
 
-    if (!this.next.isPositioned) {
-      this.next.shiftY(this.getY() + this.height + (this.chart.options.symbols[this.symbolType]['line-length'] || this.chart.options['line-length']));
-      this.next.setX(bottomPoint.x - this.next.width/2);
-      this.next.isPositioned = true;
+    var lineLength = this.chart.options.symbols[this.symbolType]['line-length'] || this.chart.options['line-length'];
 
-      this.next.render();
+    if (this.next_direction === 'right') {
+
+      var rightPoint = this.getRight();
+      var leftPoint = this.next.getLeft();
+
+      if (!this.next.isPositioned) {
+
+        this.next.setY(rightPoint.y - this.next.height/2);
+        this.next.shiftX(this.group.getBBox().x + this.width + lineLength);
+
+        var self = this;
+        (function shift() {
+          var hasSymbolUnder = false;
+          var symb;
+          for (var i = 0, len = self.chart.symbols.length; i < len; i++) {
+            symb = self.chart.symbols[i];
+
+            var diff = Math.abs(symb.getCenter().x - self.next.getCenter().x);
+            if (symb.getCenter().y > self.next.getCenter().y && diff <= self.next.width/2) {
+              hasSymbolUnder = true;
+              break;
+            }
+          }
+
+          if (hasSymbolUnder) {
+            self.next.setX(symb.getX() + symb.width + lineLength);
+            shift();
+          }
+        })();
+
+        this.next.isPositioned = true;
+
+        this.next.render();
+      }
+    } else {
+      var bottomPoint = this.getBottom();
+      var topPoint = this.next.getTop();
+
+      if (!this.next.isPositioned) {
+        this.next.shiftY(this.getY() + this.height + lineLength);
+        this.next.setX(bottomPoint.x - this.next.width/2);
+        this.next.isPositioned = true;
+
+        this.next.render();
+      }
     }
   }
 };
 
 Symbol.prototype.renderLines = function() {
   if (this.next) {
-    this.drawLineTo(this.next);
+    if (this.next_direction) {
+      this.drawLineTo(this.next, '', this.next_direction);
+    } else {
+      this.drawLineTo(this.next);
+    }
   }
 };
 
@@ -242,14 +299,12 @@ Symbol.prototype.drawLineTo = function(symbol, text, origin) {
     maxX = right.x + lineLength/2;
   } else if ((origin && origin === 'right') && isRight) {
     line = drawLine(this.chart, right, [
-      {x: symbolRight.x + lineLength/2, y: right.y},
-      {x: symbolRight.x + lineLength/2, y: symbolTop.y - lineLength/2},
-      {x: symbolTop.x, y: symbolTop.y - lineLength/2},
+      {x: symbolTop.x, y: right.y},
       {x: symbolTop.x, y: symbolTop.y}
     ], text);
     this.rightStart = true;
     symbol.topEnd = true;
-    maxX = symbolRight.x + lineLength/2;
+    maxX = right.x + lineLength/2;
   } else if ((origin && origin === 'bottom') && isOnSameColumn && isUpper) {
     line = drawLine(this.chart, bottom, [
       {x: bottom.x, y: bottom.y + lineLength/2},
@@ -261,6 +316,16 @@ Symbol.prototype.drawLineTo = function(symbol, text, origin) {
     this.bottomStart = true;
     symbol.topEnd = true;
     maxX = bottom.x + lineLength/2;
+  } else if ((origin === 'left')) {
+    line = drawLine(this.chart, left, [
+      {x: symbolTop.x + (left.x - symbolTop.x)/ 2, y: left.y},
+      {x: symbolTop.x + (left.x - symbolTop.x)/ 2, y: symbolTop.y - lineLength/2},
+      {x: symbolTop.x, y: symbolTop.y - lineLength/2},
+      {x: symbolTop.x, y: symbolTop.y}
+    ], text);
+    this.leftStart = true;
+    symbol.topEnd = true;
+    maxX = left.x;
   }
 
   if (line) {